DooT/requirements.txt
# Django and Core
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9
redis==5.0.1

# API Documentation
drf-yasg==1.21.7

# CORS and Security
django-cors-headers==4.3.1
django-filter==23.5

# Image Processing
Pillow==10.1.0

# Task Queue
celery==5.3.4

# Search
django-elasticsearch-dsl==8.0

# Environment Variables
python-decouple==3.8

# Development Tools
django-debug-toolbar==4.2.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring
sentry-sdk==1.38.0

# Payment Gateways (optional)
stripe==7.8.0
paypal-checkout-serversdk==1.0.1

# Email
django-anymail==10.2

# File Storage
boto3==1.34.0

# Caching
django-redis==5.4.0

# Validation
django-phonenumber-field==7.3.0
phonenumbers==8.13.25

# Date/Time
python-dateutil==2.8.2

# HTTP Requests
requests==2.31.0

# JSON
simplejson==3.19.2

# UUID
shortuuid==1.0.11
